Add files via upload
First iteration of a python script that runs 2 nmap scans, 1 quick common port scan, 1 long full scan. The scans are placed in a newly created directory labeled as nmap. Once the first nmap scan is completed, the script parses the data to review whether port 80 or 443 are open. If either port is open, there will be a basic dirb scan on the IP address. If neither port is open, the script will continue with the long full scan.